% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements.R
\name{extract_post_elements}
\alias{extract_post_elements}
\alias{extract_many_posts_elements}
\alias{extract_post_author_infos}
\alias{extract_many_posts_author_infos}
\alias{extract_post_text}
\alias{extract_many_posts_text}
\alias{extract_post_created_at}
\alias{extract_many_posts_created_at}
\alias{extract_post_langs}
\alias{extract_many_posts_langs}
\alias{extract_post_hashtags}
\alias{extract_many_posts_hashtags}
\alias{extract_post_hashtags_from_text}
\alias{extract_many_posts_hashtags_from_text}
\alias{create_post_url}
\alias{extract_many_posts_url}
\alias{create_post_id}
\alias{extract_many_posts_id}
\title{Extract elements from a post}
\usage{
extract_post_elements(post)

extract_many_posts_elements(posts)

extract_post_author_infos(post)

extract_many_posts_author_infos(posts)

extract_post_text(post)

extract_many_posts_text(posts)

extract_post_created_at(post)

extract_many_posts_created_at(posts)

extract_post_langs(post)

extract_many_posts_langs(posts)

extract_post_hashtags(post)

extract_many_posts_hashtags(posts)

extract_post_hashtags_from_text(text)

extract_many_posts_hashtags_from_text(posts)

create_post_url(post)

extract_many_posts_url(posts)

create_post_id(post)

extract_many_posts_id(posts)
}
\arguments{
\item{post}{A post object}

\item{posts}{A list of post objects}

\item{text}{A text}
}
\value{
A list of elements

A list of lists of elements

A list with author handle, did, and created at

A list of lists with author handle, did, and created at

The text of the post

A list of texts

The created at of the post

A list of created ats

The langs of the post

A list of langs

A list of hashtags

A list of lists of hashtags

A list of hashtags

A list of lists of hashtags

The URL of the post

A list of URLs

The ID of the post

A list of IDs
}
\description{
Extract elements from a post

Extract elements from many posts

Extract author information from a post

Extract author information from many posts

Extract text from a post

Extract text from many posts

Extract created at from a post

Extract created at from many posts

Extract langs from a post

Extract langs from many posts

Extract hashtags from a post

Extract hashtags from many posts

Extract hashtags from text

Extract hashtags from many posts

Create post URL

Create post URL from many posts

Create post ID

Create post ID from many posts
}
