% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elements_bskyr.R
\name{bsky_extract_many_posts_elements_bskyr}
\alias{bsky_extract_many_posts_elements_bskyr}
\alias{bsky_extract_many_posts_author_infos_bskyr}
\alias{bsky_extract_many_posts_text_bskyr}
\alias{bsky_extract_many_posts_created_at_bskyr}
\alias{bsky_extract_many_posts_langs_bskyr}
\alias{bsky_extract_many_posts_hashtags_bskyr}
\alias{bsky_create_post_url_bskyr}
\alias{bsky_extract_many_posts_url_bskyr}
\alias{bsky_extract_many_posts_id_bskyr}
\title{Extract elements from many posts}
\usage{
bsky_extract_many_posts_elements_bskyr(posts)

bsky_extract_many_posts_author_infos_bskyr(posts)

bsky_extract_many_posts_text_bskyr(posts)

bsky_extract_many_posts_created_at_bskyr(posts)

bsky_extract_many_posts_langs_bskyr(posts)

bsky_extract_many_posts_hashtags_bskyr(posts)

bsky_create_post_url_bskyr(post_uri, author_did)

bsky_extract_many_posts_url_bskyr(posts)

bsky_extract_many_posts_id_bskyr(posts)
}
\arguments{
\item{posts}{A list of post objects}

\item{post_uri}{The URI of the post}

\item{author_did}{The DID of the author}
}
\value{
A list of lists of elements

A list of lists with author handle, did, and created at

A list of texts

A list of created ats

A list of langs

A list of lists of hashtags

The URL of the post

A list of URLs

A list of IDs
}
\description{
Extract elements from many posts

Extract author information from many posts

Extract text from many posts

Extract created at from many posts

Extract langs from many posts

Extract hashtags from many posts

Create post URL

Create post URL from many posts

Create post ID from many posts
}
