% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Post the post}
\usage{
get_request(session, bsky_post)
}
\arguments{
\item{session}{A list containing session details}

\item{bsky_post}{The post data to be sent to BlueSky}
}
\description{
Post the post
}
