% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_bsky.R
\name{post_bsky}
\alias{post_bsky}
\title{Function to post to BlueSky}
\usage{
post_bsky(bsky_text)
}
\arguments{
\item{bsky_text}{(character) text of the BlueSky post}
}
\description{
Function to post to BlueSky
}
\examples{
\dontrun{

post_bsky("This is my first blueskyr post")
}
}
