% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_posts.R
\name{search_posts}
\alias{search_posts}
\title{Search posts for a given keyword}
\usage{
search_posts(
  keyword,
  access_jwt,
  max_posts = NULL,
  search_url = "https://bsky.social/xrpc/app.bsky.feed.searchPosts"
)
}
\arguments{
\item{keyword}{Keyword to search for}

\item{access_jwt}{Access token}

\item{max_posts}{Maximum number of posts to retrieve}

\item{search_url}{Search URL}
}
\value{
List of posts
}
\description{
Search posts for a given keyword
}
\details{
You can find more information about the Bluesky API here: \url{https://docs.bsky.app/docs/api/app-bsky-feed-search-posts}
Additionnaly more information about the "post" object is available here: \url{https://atproto.blue/en/latest/atproto/atproto_client.models.app.bsky.feed.defs.html#atproto_client.models.app.bsky.feed.defs.PostView}
}
