% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsky-clean-results.R
\name{bsky_extract_post_elements}
\alias{bsky_extract_post_elements}
\alias{bsky_extract_many_posts_elements}
\alias{bsky_extract_post_author_infos}
\alias{bsky_extract_many_posts_author_infos}
\alias{bsky_extract_post_text}
\alias{bsky_extract_many_posts_text}
\alias{extract_post_created_at}
\alias{bsky_extract_many_posts_created_at}
\alias{bsky_extract_post_langs}
\alias{bsky_extract_many_posts_langs}
\alias{bsky_extract_post_hashtags}
\alias{bsky_extract_many_posts_hashtags}
\alias{bsky_create_post_url}
\alias{bsky_extract_many_posts_url}
\alias{bsky_create_post_id}
\alias{bsky_extract_many_posts_id}
\title{Extract elements from a post}
\usage{
bsky_extract_post_elements(post)

bsky_extract_many_posts_elements(posts)

bsky_extract_post_author_infos(post)

bsky_extract_many_posts_author_infos(posts)

bsky_extract_post_text(post)

bsky_extract_many_posts_text(posts)

extract_post_created_at(post)

bsky_extract_many_posts_created_at(posts)

bsky_extract_post_langs(post)

bsky_extract_many_posts_langs(posts)

bsky_extract_post_hashtags(post)

bsky_extract_many_posts_hashtags(posts)

bsky_create_post_url(post)

bsky_extract_many_posts_url(posts)

bsky_create_post_id(post)

bsky_extract_many_posts_id(posts)
}
\arguments{
\item{post}{A post object}

\item{posts}{A list of post objects}
}
\value{
A list of elements

A list of lists of elements

A list with author handle, did, and created at

A list of lists with author handle, did, and created at

The text of the post

A list of texts

The created at of the post

A list of created ats

The langs of the post

A list of langs

A list of hashtags

A list of lists of hashtags

The URL of the post

A list of URLs

The ID of the post

A list of IDs
}
\description{
Extract elements from a post

Extract elements from many posts

Extract author information from a post

Extract author information from many posts

Extract text from a post

Extract text from many posts

Extract created at from a post

Extract created at from many posts

Extract langs from a post

Extract langs from many posts

Extract hashtags from a post

Extract hashtags from many posts

Create post URL

Create post URL from many posts

Create post ID

Create post ID from many posts
}
