% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_bearer_token}
\alias{get_bearer_token}
\title{Get bearer token}
\usage{
get_bearer_token(
  handle = Sys.getenv("bluesky_id"),
  password = Sys.getenv("bluesky_pwd"),
  login_url = "https://bsky.social/xrpc/com.atproto.server.createSession"
)
}
\arguments{
\item{handle}{Bluesky handle}

\item{password}{Bluesky password}

\item{login_url}{Bluesky login URL}
}
\value{
List with access_jwt and did
}
\description{
Get bearer token
}
\examples{
\dontrun{
get_bearer_token()
}

}
