% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsky-session.R
\name{bsky_create_session}
\alias{bsky_create_session}
\alias{bsky_get_token}
\title{Get bearer token}
\usage{
bsky_create_session(
  handle = Sys.getenv("bluesky_id"),
  password = Sys.getenv("bluesky_pwd"),
  login_url = "https://bsky.social/xrpc/com.atproto.server.createSession"
)

bsky_get_token()
}
\arguments{
\item{handle}{Bluesky handle}

\item{password}{Bluesky password}

\item{login_url}{Bluesky login URL}
}
\value{
List with access_jwt and did

Token
}
\description{
Get bearer token

Get token
}
\details{
You can find more information about the Bluesky API here: \url{https://docs.bsky.app/docs/api/com-atproto-server-create-session}
}
\examples{
\dontrun{
bsky_create_session()
}

}
