% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{create_session}
\alias{create_session}
\alias{refresh_session}
\alias{get_token}
\title{Get bearer token}
\usage{
create_session(
  handle = Sys.getenv("bluesky_id"),
  password = Sys.getenv("bluesky_pwd"),
  login_url = "https://bsky.social/xrpc/com.atproto.server.createSession"
)

refresh_session(
  refresh_token,
  refresh_url = "https://bsky.social/xrpc/com.atproto.server.refreshSession"
)

get_token()
}
\arguments{
\item{handle}{Bluesky handle}

\item{password}{Bluesky password}

\item{login_url}{Bluesky login URL}

\item{refresh_token}{Refresh token}

\item{refresh_url}{Bluesky refresh URL}
}
\value{
List with access_jwt and did

List with access_jwt and did

Token
}
\description{
Get bearer token

Refresh session

Get token
}
\details{
You can find more information about the Bluesky API here: \url{https://docs.bsky.app/docs/api/com-atproto-server-create-session}
}
\examples{
\dontrun{
create_session()
}

}
